+++ jolt-core/src/main/java/com/bazaarvoice/jolt/chainr/spec/ChainrEntry.java
@@ -18,13 +18,13 @@ package com.bazaarvoice.jolt.chainr.spec;
 import com.bazaarvoice.jolt.CardinalityTransform;
 import com.bazaarvoice.jolt.Chainr;
 import com.bazaarvoice.jolt.Defaultr;
+import com.bazaarvoice.jolt.Dictionaryr;
 import com.bazaarvoice.jolt.JoltTransform;
 import com.bazaarvoice.jolt.Modifier;
 import com.bazaarvoice.jolt.Removr;
 import com.bazaarvoice.jolt.Shiftr;
 import com.bazaarvoice.jolt.Sortr;
 import com.bazaarvoice.jolt.SpecDriven;
-import com.bazaarvoice.jolt.dictionary.spec.DictionaryDef;
 import com.bazaarvoice.jolt.exception.SpecException;
 import com.bazaarvoice.jolt.utils.StringTools;
 
@@ -57,7 +57,7 @@ public class ChainrEntry {
         temp.put( "remove", Removr.class.getName() );
         temp.put( "sort", Sortr.class.getName() );
         temp.put( "cardinality", CardinalityTransform.class.getName() );
-        temp.put("dictionaries", DictionaryDef.class.getName());
+        temp.put("dictionaries", Dictionaryr.class.getName());
         STOCK_TRANSFORMS = Collections.unmodifiableMap( temp );
     }
